+++ demo/shared/src/main/scala/games/audio/Context.scala
@@ -1,5 +1,7 @@
 package games.audio
 
+import scala.concurrent.Future
+
 import games.Resource
 import games.math.Vector3f
 import java.io.Closeable
@@ -13,9 +15,12 @@ object Format {
 }
 
 abstract class Context extends Closeable {
-  def createBufferedData(res: Resource): games.audio.Data
-  def createStreamingData(res: Resource): games.audio.Data
-  def createRawData(data: ByteBuffer, format: Format, channels: Int, freq: Int): games.audio.Data
+  def prepareStreamingData(res: Resource): Future[games.audio.Data]
+  def prepareBufferedData(res: Resource): Future[games.audio.BufferedData]
+  def prepareRawData(data: ByteBuffer, format: Format, channels: Int, freq: Int): Future[games.audio.BufferedData]
+
+  def createSource(): Source
+  def createSource3D(): Source3D
 
   def listener: Listener
 
@@ -25,10 +30,12 @@ abstract class Context extends Closeable {
   def close(): Unit = {}
 }
 
-abstract class Listener extends Closeable {
+sealed trait Spatial {
   def position: Vector3f
   def position_=(position: Vector3f)
+}
 
+abstract class Listener extends Closeable with Spatial {
   def up: Vector3f
 
   def orientation: Vector3f
@@ -37,3 +44,37 @@ abstract class Listener extends Closeable {
 
   def close(): Unit = {}
 }
+
+abstract class Data extends Closeable {
+  def attach(source: AbstractSource): Future[games.audio.Control]
+
+  def close(): Unit = {}
+}
+
+abstract class BufferedData extends Data {
+  def attachNow(source: AbstractSource): games.audio.Control
+}
+
+abstract class Control extends Closeable {
+  def play: Unit
+  def pause: Unit
+
+  def playing: Boolean
+
+  def volume: Float
+  def volume_=(volume: Float)
+
+  def loop: Boolean
+  def loop_=(loop: Boolean)
+
+  def pitch: Float
+  def pitch_=(pitch: Float)
+
+  def close(): Unit = {
+    this.pause
+  }
+}
+
+sealed abstract class AbstractSource
+abstract class Source extends AbstractSource
+abstract class Source3D extends AbstractSource with Spatial
